19.09.2020 23:49 Попытка реализовать ДЗ занятия 14

+++ src/main/java/ru/sbrf/payment/common/exchange/ClientPaymentContainer.java
@@ -6,11 +6,11 @@ import java.io.Serializable;
 @Getter
 public class ClientPaymentContainer extends Container implements Serializable {
     private String account;
-    private Integer summa;
+    private Long summa;
     private Integer currency;
     private String mobileNumber;
 
-    public ClientPaymentContainer(String clientNumber, String account, Integer summa, Integer currency, String mobileNumber) {
+    public ClientPaymentContainer(String clientNumber, String account, Long summa, Integer currency, String mobileNumber) {
         super(clientNumber);
         this.account = account;
         this.summa = summa;
